renamed to intake 4, deprecated intake 3 version, updated imports
